refactor: move logic to a Token Helper
